refactor(Redux): made some changes to where redux is being called
