+++ src/components/ImageGallery.tsx
@@ -1,31 +1,45 @@
-import React from 'react';
+import React, { useEffect } from 'react';
 import Card from './Card';
+import {
+  useAppDispatch,
+  useAppSelector,
+} from '../store/hooks';
+// import {
+//   getPhotos,
+//   selectPhotos,
+// } from '../slices/photoSlice';
 import styles from '../../styles/ImageGallery.module.css';
 
-const ImageGallery = () => (
-  <section className={styles.gallery}>
-    <Card />
-    <Card />
-    <Card />
-    <Card />
-    <Card />
-    <Card />
-    <Card />
-    <Card />
-    <Card />
-    {/* <div className={styles.card}>1</div>
-    <div className={styles.card}>2</div>
-    <div className={styles.card}>3</div>
-    <div className={styles.card}>4</div>
-    <div className={styles.card}>5</div>
-    <div className={styles.card}>6</div>
-    <div className={styles.card}>7</div>
-    <div className={styles.card}>8</div>
-    <div className={styles.card}>9</div>
-    <div className={styles.cardSm}>9</div>
-    <div className={styles.cardSm}>9</div>
-    <div className={styles.cardSm}>9</div> */}
-  </section>
-);
+const ImageGallery = () => {
+  const dispatch = useAppDispatch();
+  // const {
+  //   data,
+  //   pending,
+  //   error,
+  // } = useAppSelector(selectPhotos);
+
+  // useEffect(() => {
+  //   dispatch(getPhotos('venice'));
+  // }, [dispatch]);
+
+  // console.log(data.photos);
+
+  return (
+    <section className={styles.gallery}>
+      <Card />
+      <Card />
+      <Card />
+      <Card />
+      <Card />
+      <Card />
+      <Card />
+      <Card />
+      <Card />
+      {/* {pending && <p>Loading...</p>}
+      {data && <p>{data}</p>}
+      {error && <p>Oops, something went wrong</p>} */}
+    </section>
+  );
+};
 
 export default ImageGallery;
